% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_conversion.R
\name{lunar_to_gregorian}
\alias{lunar_to_gregorian}
\title{Lunar Calendar conversion}
\usage{
lunar_to_gregorian(lunar_date)
}
\arguments{
\item{lunar_date}{a string with the lunar date in the following format "YYYY-MM-DD "or "YYYY-MML-DD", if it is a leap month add "L" after the month MM}
}
\value{
The Gregorian date as a string
}
\description{
Convert Lunar date to Gregorian date
}
\examples{
lunar_to_gregorian("1888-12-24");
lunar_to_gregorian(c("1888-12-24", "1966-03L-12"));
}
